+++ src/test/java/org/starchartlabs/flare/test/publishing/plugin/DockerBuildPluginTest.java
@@ -46,4 +46,10 @@ public class DockerBuildPluginTest {
         Assert.assertNotNull(task);
     }
 
+    @Test
+    public void cleanContainerTaskAdded() throws Exception {
+        Task task = project.getTasks().getByName("cleanContainer");
+        Assert.assertNotNull(task);
+    }
+
 }